Refactor editor scripts related to ConeRayEstimator
